Move src/components/tabLayout/ to src/apps/mydb/elements/tabLayout/
